server/requirements.txt
requests
Django
Pillow
gunicorn
python-dotenv
whitenoise[brotli]